// packages/valnor-postgres/src/__tests__/codegen/one_sql.account-table.ts
/* File generated by Valnor. Do not edit. */
import * as valnor from "valnor";
import * as udt from "./one_sql-enums.js";

export const Account = valnor.newSqlTable({
   name: "account",
   schema: "one_sql",
   types: <{ Select: IAccountSelect, Insert: IAccountInsert, Update: IAccountUpdate }>{}, 
   pk: "account_pk", 
},{

   /**
    * first_name varchar
   */
   firstName: "first_name",

   /**
    * account_id uuid default gen_random_uuid()
   */
   accountId: "account_id",

   /**
    * status account_status default 'created'::one_sql.account_status
   */
   status: "status",

   /**
    * created_at timestamptz default now()
   */
   createdAt: "created_at",

   /**
    * modified_at timestamptz default now()
   */
   modifiedAt: "modified_at",

   /**
    * last_name varchar
   */
   lastName: "last_name",

   /**
    * notes text
   */
   notes: "notes",

   /**
    * email varchar
   */
   email: "email",
});

export type IAccountInsert = {
   firstName: string;
   accountId?: string;
   status?: udt.AccountStatusUdt;
   createdAt?: Date;
   modifiedAt?: Date;
   lastName: string;
   notes?: string | null;
   email: string;
}

export type IAccountUpdate = Partial<IAccountInsert>;

export type IAccountSelect = {
   readonly firstName: string
   readonly accountId: string
   readonly status: udt.AccountStatusUdt
   readonly createdAt: Date
   readonly modifiedAt: Date
   readonly lastName: string
   readonly notes: string | null
   readonly email: string
}

export type IAccountJson = valnor.JsonRow<IAccountSelect>;